split wave parameters into base class
